task(QRCode): Moved and considered as a Modal component
