+++ components/base/Modal/components/QRCode.tsx
