+++ components/base/Modal/index.tsx
@@ -5,3 +5,4 @@ export { default as ModalShare } from './ModalShare';
 export { default as ModalShowcase } from './ModalShowcase';
 export { default as ModalWallet } from './ModalWallet';
 export { default as ModalLoader } from './components/ModalLoader';
+export { default as QRCode } from './components/QRCode';
